requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.7
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mysql-connector-python==9.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.2
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
